100-prime_factor.c modified

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -9,7 +9,7 @@
 int main(void)
 {
 	long n = 612852475143;
-	long m;
+	int m;
 
 	while (m <= n / 2)
 	{
